Database change to authentication
