+++ apps/authentication/admin.py
@@ -5,7 +5,17 @@
 
 from django.contrib import admin
 from .models import Profile
+from .models import Plant, Farm, House, Line, Pole, LED
 
 # Register your models here.
 # admin.site.register(User)
 admin.site.register(Profile)
+
+
+# Register your models here.
+admin.site.register(Plant)
+admin.site.register(Farm)
+admin.site.register(House)
+admin.site.register(Line)
+admin.site.register(Pole)
+admin.site.register(LED)
